move fragments and an activity
