+++ app/src/main/AndroidManifest.xml
@@ -11,9 +11,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.AndroidEngineerCodeCheck"
-        android:fullBackupContent="@xml/backup_descriptor">
+        android:fullBackupContent="@xml/backup_descriptor"
+        android:dataExtractionRules="@xml/data_extraction_rules">
         <activity
-            android:name="jp.co.yumemi.android.codeCheck.TopActivity"
+            android:name="jp.co.yumemi.android.codeCheck.ui.TopActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
