+++ app/src/main/kotlin/jp/co/yumemi/android/codeCheck/MainApplication.kt
@@ -1,0 +1,4 @@
+package jp.co.yumemi.android.codeCheck
+
+class MainApplication {
+}
