+++ app/src/main/kotlin/jp/co/yumemi/android/codeCheck/ui/CustomAdapter.kt
@@ -4,11 +4,21 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import jp.co.yumemi.android.codeCheck.domain.Item
 import jp.co.yumemi.android.codeCheck.R
-import jp.co.yumemi.android.codeCheck.diff_util
+
+val diff_util = object : DiffUtil.ItemCallback<Item>() {
+    override fun areItemsTheSame(oldItem: Item, newItem: Item): Boolean {
+        return oldItem.name == newItem.name
+    }
+
+    override fun areContentsTheSame(oldItem: Item, newItem: Item): Boolean {
+        return oldItem == newItem
+    }
+}
 
 class CustomAdapter(
     private val itemClickAction: ItemClickAction,
