+++ app/src/main/kotlin/jp/co/yumemi/android/codeCheck/ui/TopActivity.kt
@@ -1,10 +1,11 @@
 /*
  * Copyright © 2021 YUMEMI Inc. All rights reserved.
  */
-package jp.co.yumemi.android.codeCheck
+package jp.co.yumemi.android.codeCheck.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import java.util.Date
+import jp.co.yumemi.android.codeCheck.R
 
 class TopActivity : AppCompatActivity(R.layout.activity_top) {
